+++ extension/webview/src/components/FileMentionDropdown/index.ts
@@ -1,0 +1,1 @@
+export { FileMentionDropdown } from './FileMentionDropdown';